applications: serial_lte_modem: Minor coverity fixes
Fixed one instance of usage of 32-bit time.
Fixed one instance of uninitialized scalar where the
buffer might have been used uninitialized.

+++ applications/serial_lte_modem/src/slm_at_socket.c
@@ -394,7 +394,7 @@ static int do_socketopt_get(int option)
 		if (ret) {
 			LOG_ERR("getsockopt(%d) error: %d", option, -errno);
 		} else {
-			rsp_send("\r\n#XSOCKETOPT: \"%d sec\"\r\n", (int)tmo.tv_sec);
+			rsp_send("\r\n#XSOCKETOPT: \"%ld sec\"\r\n", (long)tmo.tv_sec);
 		}
 	} break;
